Only ignore relpaths, not entire paths
There was a problem when my project repo got checked out by a CI runner, where
it got placed into a '.../build/myproject' folder.

In .gitignore, there was a '/BUILD/' ignore path.

As a result, fnmatch.fnmatch(path, '*BUILD*') in BaseWorkingCopy.ignored
returned True (as I understand, it is case insensitive), eventually
resulting in my entire tree being ignored when searching for refs.

There was a 'fix' for this, which don't ignore '*build*' if the
environment variable 'CI' was set.  This didn't match my '*BUILD*'
pattern, so it didn't trigger.  Also, it would actually not ignore a
real 'build/' directory in my repository if there was one, so I consider
this a bug anyway.