Add --all option to `doorstep edit` and change default behaviour of this cmd.
Now, `doorstep edit ITEM` only the text part of the item by default.
To edit the whole item with all of its attributes, one must add the
--all or -a option to the command.

To edit the text part only, a temporary file is created to be edited
by the user, and then read by doorstop to update the actual item file.