Remove superfluous @auto_load
After a item save() there is no need to load the item, so set _loaded to
True in this function.  The item setters do not need to load the item.
Remove an implicit load() in save() through the data property.  No
changes in the test suite are necessary due to this code change.