ItemValidator class to encapsulate validations of Item
The Item class is very close to hitting 1000 lines limit which indicates that
it probably does too many things.

As discussed in https://github.com/doorstop-dev/doorstop/issues/409, which
identified some possible targets for refactoring, the validation aspects
could be moved to a separate module.

This changeset moves all 4 validation-related functions to a separate class:
`ItemValidator`.

With this change, `Item` is no longer `BaseValidatable` which seems reasonable,
because otherwise it would be hard to move these 4 functions out without
introducing cyclic dependencies.