Improve fingerprinting of extended attributes
Do not issue a warning in case a reviewed attribute is missing.  The
doorstop tool has not enough information to do a proper validation of
items of extended attributes.

Make the value serialization independent of the YAML file format.  Use a
simple serialization based on Python string representations.  This helps
in case the item storage format changes in the future.

This is a breaking change, since the fingerprint of existing items may
change.